Refactor ServiceNow integration and enhance lab management functionality
- Replaced `ServiceNowHandler` with `ServiceNowAutoAssign` in `main.py` to streamline ticket assignment processes.
- Introduced `JiraHandler` for improved Jira login and ticket management within the `SnowAIProcessor`.
- Added a new `JiraHandler` class to handle Jira authentication and operations, enhancing modularity.
- Implemented waiting mechanisms in `LabManager` to ensure lab status checks are reliable during operations.
- Created a new `jira_handler.py` module for dedicated Jira operations, separating concerns for better maintainability.
- Updated `ConfigManager` to disable interpolation for URLs, improving configuration handling.
- Enhanced error handling and logging throughout the integration processes for better debugging and user feedback.
