+++ lx_toolbox/core/__init__.py
@@ -4,10 +4,14 @@ from .base_selenium_driver import BaseSeleniumDriver
 from .lab_manager import LabManager
 from .link_checker import LinkChecker
 from .jira_handler import JiraHandler
+from .servicenow_handler import ServiceNowHandler
+from .servicenow_autoassign import ServiceNowAutoAssign
 
 __all__ = [
     'BaseSeleniumDriver',
     'LabManager', 
     'LinkChecker',
     'JiraHandler',
+    'ServiceNowHandler',
+    'ServiceNowAutoAssign',
 ]
